requirements.txt
et-xmlfile==1.1.0
icalendar==5.0.11
openpyxl==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
